pyproject.toml
[tool.poetry]
name = "pesquisa"
version = "0.1.0"
description = ""
authors = ["Joelson Sartori Junior <33182102+JoelsonSartoriJr@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "0.28"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
wordcloud = "^1.9.3"
pypdf2 = "^3.0.1"
serpapi = "^0.1.5"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.1"
scholarly = "^1.7.11"
sentence-transformers = "^3.1.1"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
